# backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: Matchlock Backend API
  version: 0.1.0
  description: |
    REST API served by `backend-go` (keeper process). Match data is ingested from TxLINE SSE
    and cached in Redis. Wager accounts are read from Solana RPC.

    **CORS:** `GET` and `OPTIONS` are allowed from origins in `CORS_ALLOWED_ORIGINS`
    (default `http://localhost:5173`, `http://127.0.0.1:5173`).

    **On-chain writes** (`make_wager`, `accept_wager`, `cancel_wager`) are signed in the
    browser wallet via the Solana program — this API is read-only for on-chain state.

    **Devnet defaults** (see `backend-go/.env`):
    - Program: `B39Vk22T2VPpqBEbGkW51BzFC6sNeeiQQ1mqwdvCJ2H4`
    - Stablecoin mint: `ELWTKspHKCnCfCiCiqYw1EDH77k8VCP74dK9qytG2Ujh`
    - Cluster: `devnet`
servers:
  - url: http://localhost:8080
    description: Local keeper (`HTTP_ADDR`, default `:8080`)
  - url: http://127.0.0.1:8080
    description: Local keeper (loopback)
...
paths:
...
  /matches:
    get:
      tags: [matches]
      operationId: listMatches
      summary: List cached matches
      description: |
        Returns all matches currently in the Redis cache, populated by the TxLINE scores
        SSE ingestor. Empty array when no fixtures have been observed yet.
      responses:
        "200":
          description: Match list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchView"
        "500":
          $ref: "#/components/responses/CacheError"
...
  /matches/{id}:
    get:
      tags: [matches]
      operationId: getMatch
      summary: Get match by ID
      parameters:
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: Match found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchView"
        "400":
          $ref: "#/components/responses/InvalidMatchId"
        "404":
          $ref: "#/components/responses/MatchNotFound"
        "500":
          $ref: "#/components/responses/CacheError"
  /wagers:
    get:
      tags: [wagers]
      operationId: listWagers
      summary: List on-chain wagers
      description: |
        Scans Matchlock program accounts via Solana RPC with optional filters.

        When no `status` or `wallet` filter is supplied, only `open` and `matched` wagers
        are returned (market browsing). Use `wallet` to query a user's full history.
      parameters:
        - $ref: "#/components/parameters/MatchIdQuery"
        - $ref: "#/components/parameters/StatusQuery"
        - $ref: "#/components/parameters/WalletQuery"
      responses:
        "200":
          description: Wager list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WagerView"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "502":
          $ref: "#/components/responses/RpcError"
...
  /wagers/{pubkey}:
    get:
      tags: [wagers]
      operationId: getWager
      summary: Get wager by account pubkey
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "200":
          description: Wager found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WagerView"
        "400":
          $ref: "#/components/responses/InvalidPubkey"
        "404":
          $ref: "#/components/responses/WagerNotFound"
        "502":
          $ref: "#/components/responses/RpcError"
  /wagers/{pubkey}/settlement:
    get:
      tags: [settlement]
      operationId: getWagerSettlement
      summary: Get settlement status for a matched wager
      description: |
        Returns the current settlement state of a matched wager. States include:
        `match_live`, `match_ended_unverified`, `claimable`, `refundable`, `queued`,
        `retrying`, `settled`, `failed`, and `not_applicable`.
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "200":
          description: Settlement status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementStatusView"
        "400":
          $ref: "#/components/responses/InvalidPubkey"
        "502":
          $ref: "#/components/responses/RpcError"
...
components:
  parameters:
    MatchId:
      name: id
      in: path
      required: true
      description: Match identifier (decimal fixture id string)
      schema:
        type: string
        example: "17952170"
    MatchIdQuery:
      name: match_id
      in: query
      required: false
      description: Filter by match fixture id
      schema:
        type: string
        example: "17952170"
    WagerPubkey:
      name: pubkey
      in: path
      required: true
      description: Base58-encoded Solana account pubkey
      schema:
        type: string
        example: "8fwJw3XMgFgvjar6oMY72zmfE1K1ycJhQRhSL5LEJMjK"
...
    StatusQuery:
      name: status
      in: query
      required: false
      description: Filter by wager status
      schema:
        $ref: "#/components/schemas/WagerStatus"
    WalletQuery:
      name: wallet
      in: query
      required: false
      description: Filter by wallet pubkey (returns full history for that wallet)
      schema:
        type: string
        example: "2qtn5WqZhXxEpgVqJy1u2m9RjXDkWsP2NdGq9BmPQGRP"
...
  schemas:
...
    MatchView:
      type: object
      required:
        - match_id
        - fixture_id
        - status
        - is_final
        - seq
        - updated_at
        - status_stale
      properties:
        match_id:
          type: string
          description: Canonical id (decimal fixture id)
        fixture_id:
          type: integer
          format: int64
        status:
          type: string
          description: TxLINE gameState (e.g. HT, F2, scheduled)
        is_final:
          type: boolean
        final_source:
          type: string
          description: How finality was determined (e.g. txline, inferred)
        home_goals:
          type: integer
          format: int32
          nullable: true
        away_goals:
          type: integer
          format: int32
          nullable: true
        seq:
          type: integer
          format: int32
          description: Latest observed TxLINE sequence number
        updated_at:
          type: string
          format: date-time
        finalized_at:
          type: string
          format: date-time
          nullable: true
        status_stale:
          type: boolean
          description: True when a non-final fixture has exceeded the maximum live-status age
        start_time:
          type: integer
          format: int64
          description: Kickoff epoch ms
        competition_id:
          type: integer
          format: int32
        competition:
          type: string
        fixture_group_id:
          type: integer
          format: int32
        participant1_id:
          type: integer
          format: int32
        participant2_id:
          type: integer
          format: int32
        participant1_is_home:
          type: boolean
          description: Whether TxLINE participant 1 maps to the home team.
        home_team:
          type: string
        away_team:
          type: string
        sport_id:
          type: integer
          format: int32
        country_id:
          type: integer
          format: int32
        odds:
          allOf:
            - $ref: "#/components/schemas/MatchOddsView"
          nullable: true
    MatchOddsView:
      type: object
      properties:
        home:
          type: number
          format: double
        draw:
          type: number
          format: double
        away:
          type: number
          format: double

    # ── Wagers ──
    WagerView:
      type: object
      required:
        - pubkey
        - maker
        - taker
        - match_id
        - maker_side
        - stake
        - status
      properties:
        pubkey:
          type: string
          description: Wager account address (base58)
        maker:
          type: string
          description: Maker wallet (base58)
        invited_taker:
          type: string
          description: Designated taker wallet (base58), if any
        taker:
          type: string
          description: Taker wallet (base58); zero key when open
        match_id:
          type: string
          description: Fixture id string
        maker_side:
          $ref: "#/components/schemas/Side"
        taker_side:
          $ref: "#/components/schemas/Side"
        stake:
          type: integer
          format: int64
          description: Stake per side in stablecoin smallest units (6 decimals on devnet USDT)
          example: 100000
        status:
          $ref: "#/components/schemas/WagerStatus"
    Side:
      type: string
      enum: [home, away]
    WagerStatus:
      type: string
      enum: [open, matched, settled, cancelled]
...
    SettlementStatusView:
      type: object
      required:
        - state
        - message
        - match_final
        - updated_at
      properties:
        state:
          type: string
          enum:
            - match_live
            - match_ended_unverified
            - claimable
            - refundable
            - queued
            - retrying
            - settled
            - failed
            - not_applicable
        message:
          type: string
          description: User-facing status message
        match_final:
          type: boolean
        settled_at:
          type: string
          format: date-time
          nullable: true
        tx_signature:
          type: string
          description: On-chain settlement transaction signature
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message
        code:
          type: string
          description: Machine-readable error code
  responses:
    CacheError:
      description: Redis cache error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: failed to list matches
            code: CACHE_ERROR
    RpcError:
      description: Solana RPC error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: failed to list wagers
            code: RPC_ERROR
    InvalidMatchId:
      description: Missing or empty match id
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: match id is required
            code: INVALID_MATCH_ID
    MatchNotFound:
      description: Match not in cache
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: match not found
            code: MATCH_NOT_FOUND
    InvalidPubkey:
      description: Invalid or missing wager pubkey
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: invalid wager pubkey
            code: INVALID_PUBKEY
    WagerNotFound:
      description: Wager account not found on chain
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: wager not found
            code: WAGER_NOT_FOUND
    InvalidQuery:
      description: Invalid query parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: status must be open, matched, settled, or cancelled
            code: INVALID_QUERY
...